Create index.html
This HTML code represents the structure of a simple weather app. Here's a breakdown of its components:

Head Section:

Character set and viewport settings.
Title of the web page.
Link to an external stylesheet (style.css) for styling.
Link to the Boxicons library for icon fonts.
External script reference to script.js.
Body Section:

The main container div with the class container.
Inside the container, there are:
A search box with an input field, a location icon (bxs-map), and a search icon (bx-search-alt-2).
A div with the class city-hide.
Another div with the class weather-box containing weather information.
Inside weather-box, there is a div with the class box.
Inside the box, there is a div with the class info-weather.
Inside info-weather, there is a div with the class weather.
Inside weather, there is an image (cloud1.png), temperature (temperature), and weather description (description).
Weather details including humidity and wind information.
Humidity is represented by an icon (bx-water) and percentage value.
Wind speed is represented by an icon (bx-wind) and speed value in Km/h.
A div with the class not-found for displaying a message when the location is not found.
Inside not-found, there is a div with the class box containing an image (404.png) and a message.
Scripts:

An external script file (script.js) is referenced at the end of the body section.
This structure suggests a weather app with a search functionality, displaying current weather information, and providing details such as temperature, humidity, and wind speed.